Corrected region offset value
Since we assume for now that the memory regions are aligned by
pagesize, the mem.offs value is 0.

The VHOST_VIRTIO_ADD_REGION and VHOST_VIRTIO_GET_REGION still
return offset value. The offset value is calculated as
<memory region index> * PAGE_SIZE.

+++ virtio-uio/files/uio-virtio.h
@@ -20,8 +20,8 @@
 struct virtio_region {
         uint32_t indx;
         uint32_t offs;
-        uint32_t addr;
-        uint32_t size;
+        uint64_t addr;
+        uint64_t size;
 };
 
 #define VHOST_VIRTIO_ADD_REGION _IOWR(VHOST_VIRTIO, 0x90, struct virtio_region)
